requirements-dev.txt
tox==3.24.4
pytest==6.2.5
pytest-cov==3.0.0
pytest-mock==3.7.0
pytest-aiohttp